Remove space roles before org role

+++ cloudfoundry/strip-user-org-and-space-roles.sh
@@ -3,6 +3,9 @@
 set -euo pipefail
 shopt -s inherit_errexit || true
 
+MAX_SPACE_ROLES=100
+MAX_ORG_ROLES=20
+
 main() {
   [[ $# -eq 2 ]] || usage "Expected two arguments, got $#"
 
@@ -16,10 +19,29 @@ main() {
     exit 1
   fi
 
-  ORGANIZATION_GUID=$(cf org "$org" --guid)
+  if ! ORGANIZATION_GUID=$(cf org "$org" --guid); then
+    echo "no org found for $org. Either it doesn't exist or something else went wrong"
+    exit 1
+  fi
 
-  # get all user roles for the org, including organization_user
-  for role_guid in $(cf curl "/v3/roles?user_guids=$USER_GUID&organization_guids=$ORGANIZATION_GUID&per_page=5000" | jq -r '.resources[].guid'); do
+  # get all space guids belonging to the org
+  SPACE_GUIDS=$(cf curl "/v3/spaces?organization_guids=$ORGANIZATION_GUID" | jq -r '.resources[].guid')
+
+  if [[ -n "$SPACE_GUIDS" ]]; then
+    local space_guids_csv
+    space_guids_csv=$(echo "$SPACE_GUIDS" | paste -sd, -)
+
+    for role_guid in $(cf curl "/v3/roles?user_guids=$USER_GUID&space_guids=$space_guids_csv&per_page=$MAX_SPACE_ROLES" | jq -r '.resources[].guid'); do
+      echo removing space role with: cf curl -X DELETE "/v3/roles/$role_guid"
+      cf curl -X DELETE "/v3/roles/$role_guid"
+    done
+  fi
+
+  # now remove all org-level roles for the user, including organization_user
+  for role_guid in $(cf curl "/v3/roles?user_guids=$USER_GUID&organization_guids=$ORGANIZATION_GUID&per_page=$MAX_ORG_ROLES" | jq -r '.resources[].guid'); do
+    echo removing org role with: cf curl -X DELETE "/v3/roles/$role_guid"
+    # Let the above role removals propagate (sleep 1 was insufficent)
+    sleep 5
     cf curl -X DELETE "/v3/roles/$role_guid"
   done
 }
@@ -27,13 +49,13 @@ main() {
 usage() {
   [[ $# -gt 0 ]] && echo "ERROR: $*"
   cat <<EOF
-  USAGE: $(basename "$0") USER ORG SPACE
+  USAGE: $(basename "$0") USER ORG
 
   Removes org and space roles for user.
 
   Examples:
 
-    $(basename "$0") bob accounting dev
+    $(basename "$0") bob@cfo.gov accounting
 EOF
   exit 1
 }
